Move `shape::Geometry::transform` to `Shape`
