+++ crates/fj-kernel/src/shape/geometry.rs
@@ -1,12 +1,9 @@
-use fj_math::{Point, Transform};
+use fj_math::Point;
 
-use crate::{
-    geometry::{Curve, Surface},
-    topology::Face,
-};
+use crate::geometry::{Curve, Surface};
 
 use super::{
-    stores::{Curves, Faces, Points, Surfaces},
+    stores::{Curves, Points, Surfaces},
     Iter,
 };
 
@@ -15,42 +12,9 @@ pub struct Geometry<'r> {
     pub(super) points: &'r mut Points,
     pub(super) curves: &'r mut Curves,
     pub(super) surfaces: &'r mut Surfaces,
-
-    // This is needed here for a weird workaround, which in turn is necessary
-    // because triangle representation still exists. Once triangle
-    // representation is no longer a thing, this field can be moved to
-    // `Topology`, where it belongs.
-    //
-    // This issue has some context on triangle representation:
-    // https://github.com/hannobraun/Fornjot/issues/97
-    pub(super) faces: &'r mut Faces,
 }
 
 impl Geometry<'_> {
-    /// Transform the geometry of the shape
-    ///
-    /// Since the topological types refer to geometry, and don't contain any
-    /// geometry themselves, this transforms the whole shape.
-    pub fn transform(&mut self, transform: &Transform) {
-        self.points
-            .update(|point| *point = transform.transform_point(point));
-        self.curves
-            .update(|curve| *curve = curve.transform(transform));
-        self.surfaces
-            .update(|surface| *surface = surface.transform(transform));
-
-        // While some faces use triangle representation, we need this weird
-        // workaround here.
-        self.faces.update(|mut face| {
-            use std::ops::DerefMut as _;
-            if let Face::Triangles(triangles) = face.deref_mut() {
-                for (triangle, _) in triangles {
-                    *triangle = transform.transform_triangle(triangle);
-                }
-            }
-        });
-    }
-
     /// Access an iterator over all points
     ///
     /// The caller must not make any assumptions about the order of points.
